requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
accu==0.1a0
    # via -r requirements.in
arrow==1.2.2
    # via django-q
asgiref==3.5.2
    # via django
babel==2.10.3
    # via py-moneyed
bleach[css]==5.0.1
    # via django-markdownify
blessed==1.19.1
    # via django-q
brotli==1.0.9
    # via fonttools
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   cryptography
    #   weasyprint
charset-normalizer==2.1.0
    # via requests
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==3.4.8
    # via
    #   -r requirements.in
    #   pyjwt
cssselect2==0.6.0
    # via weasyprint
defusedxml==0.7.1
    # via
    #   odfpy
    #   python3-openid
diff-match-patch==20200713
    # via django-import-export
django==3.2.15
    # via
    #   -r requirements.in
    #   accu
    #   django-allauth
    #   django-allauth-2fa
    #   django-cors-headers
    #   django-error-report
    #   django-filter
    #   django-formtools
    #   django-import-export
    #   django-js-asset
    #   django-markdownify
    #   django-money
    #   django-mptt
    #   django-otp
    #   django-picklefield
    #   django-q
    #   django-redis
    #   django-sql-utils
    #   django-sslserver
    #   django-stdimage
    #   django-user-sessions
    #   django-weasyprint
    #   django-xforwardedfor-middleware
    #   djangorestframework
django-allauth==0.51.0
    # via
    #   -r requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements.in
django-cleanup==6.0.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-error-report==0.2.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-formtools==2.3
    # via -r requirements.in
django-import-export==2.5.0
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-mptt
django-maintenance-mode==0.16.3
    # via -r requirements.in
django-markdownify==0.9.2
    # via -r requirements.in
django-money==2.1.1
    # via -r requirements.in
django-mptt==0.11.0
    # via -r requirements.in
django-otp==1.1.3
    # via django-allauth-2fa
django-picklefield==3.1
    # via django-q
django-q==1.3.9
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-sql-utils==0.6.1
    # via -r requirements.in
django-sslserver==0.22
    # via -r requirements.in
django-stdimage==5.3.0
    # via -r requirements.in
django-user-sessions==1.7.1
    # via -r requirements.in
django-weasyprint==2.1.0
    # via -r requirements.in
django-xforwardedfor-middleware==2.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fonttools[woff]==4.34.4
    # via weasyprint
gunicorn==20.1.0
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via markdown
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markdown==3.4.1
    # via django-markdownify
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.0
    # via requests-oauthlib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
pdf2image==1.16.0
    # via -r requirements.in
pillow==9.2.0
    # via
    #   -r requirements.in
    #   django-stdimage
    #   pdf2image
    #   python-barcode
    #   qrcode
    #   weasyprint
py-moneyed==1.2
    # via
    #   -r requirements.in
    #   django-money
pycparser==2.21
    # via cffi
pydyf==0.2.0
    # via weasyprint
pyjwt[crypto]==2.4.0
    # via django-allauth
pyphen==0.12.0
    # via weasyprint
python-barcode[images]==0.14.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
python-fsutil==0.6.1
    # via django-maintenance-mode
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   babel
    #   django
    #   djangorestframework
pyyaml==6.0
    # via tablib
qrcode[pil]==7.3.1
    # via
    #   -r requirements.in
    #   django-allauth-2fa
rapidfuzz==0.7.6
    # via -r requirements.in
redis==3.5.3
    # via
    #   django-q
    #   django-redis
regex==2022.8.17
    # via -r requirements.in
requests==2.28.1
    # via
    #   coreapi
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.9.0
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   blessed
    #   html5lib
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-sql-utils
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via
    #   -r requirements.in
    #   django-import-export
tinycss2==1.1.1
    # via
    #   bleach
    #   cssselect2
    #   weasyprint
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.11
    # via
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via blessed
weasyprint==54.3
    # via
    #   -r requirements.in
    #   django-weasyprint
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zipp==3.8.1
    # via importlib-metadata
zopfli==0.2.1
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
